python_code_for_multivariable_linear_regression
The following code file contains python code for predicting the housing price using multivariable linear regression, given two features, number of bedrooms and size. The ﬁle ex1data2.txt contains a training set of housing prices in Portland, Oregon. The ﬁrst column is the size of the house (in square feet), the second column is the number of bedrooms, and the third column is the price of the house.